requirements.txt
spacy
nltk
tqdm
scikit-learn
pyyaml
transformers
pandas
fasttext
gensim
setfit
datasets
frame-semantic-transformer
# install torch based on your system requirements